refactor(game): move Windows and SDL2 input impls into platform
