+++ src/game/input/platform/windows/input_dinput8.h
